5th commit

+++ controllers/todosController.js
@@ -1,19 +1,22 @@
 const Todos = require("../models/Todo");
+const httpStatusCodes= require('http-status-codes')
 
 const BadRequestError = require("../errors/bad-request");
 
 const getTodos = async (req, res) => {
-  const todos = await Todos.find({});
-  res.status(200).json({ todos, amount: todos.length , user:req.user});
+  console.log(req.user)
+  const todos = await Todos.find({createdBy:req.user.userId});
+  res.status(httpStatusCodes.OK).json({ todos, amount: todos.length , user:req.user});
 };
 
 const createTodo = async (req, res) => {
   const { name, description } = req.body;
+  
   if (!name || !description) {
     throw new BadRequestError("Please provide all values");
   }
-  const todo = await Todos.create({ name, description });
-  res.json({ todo });
+  const todo = await Todos.create({ name, description,createdBy:req.user.userId });
+  res.status(httpStatusCodes.CREATED).json({ todo });
 };
 
 const updateTodo = async (req, res) => {
@@ -23,27 +26,27 @@ const updateTodo = async (req, res) => {
   //  return res.status(400).send("please provide all values");
   //}
 
-  const task = await Todos.findOneAndUpdate({ _id: id }, req.body, {
+  const task = await Todos.findOneAndUpdate({ _id: id, createdBy:req.user.userId }, req.body, {
     new: true,
     runValidators: true,
   });
   if (!task) {
     return res
-      .status(404)
+      .status(httpStatusCodes.NOT_FOUND)
       .json({ msg: `no item with id ${id}`, success: false });
   }
-  res.status(200).json({ task });
+  res.status(httpStatusCodes.OK).json({ task });
 };
 
 const deleteTodo = async (req, res) => {
   const { id } = req.params;
-  const task = await Todos.findOneAndDelete({ _id: id });
+  const task = await Todos.findOneAndDelete({ _id: id, createdBy:req.user.userId });
   if (!task) {
     return res
-      .status(404)
+      .status(httpStatusCodes.NOT_FOUND)
       .json({ msg: `no item with id ${id}`, success: false });
   }
-  res.status(200).json({ task, msg: "successs" });
+  res.status(httpStatusCodes.OK).json({ task, msg: "successs" });
 };
 
 module.exports = {
